Clean up builders
Builder belongs in EvvoIsland.scala, not in EvvoIslandActor.scala.

Provide builder method in all ...Island... classes.
