+++ src/main/scala/com/evvo/island/EvvoIsland.scala
@@ -2,9 +2,11 @@ package com.evvo.island
 
 import java.util.Calendar
 
+import akka.actor.{ActorSystem, Props}
 import akka.event.LoggingAdapter
-import com.evvo.agent.{CreatorAgent, DeletorAgent, MutatorAgent, TCreatorFunc, TDeletorFunc, TMutatorFunc}
-import com.evvo.island.population.{Population, TObjective, TParetoFrontier}
+import com.evvo.{CreatorFunctionType, DeletorFunctionType, MutatorFunctionType, ObjectiveFunctionType}
+import com.evvo.agent.{CreatorAgent, CreatorFunc, DeletorAgent, DeletorFunc, MutatorAgent, MutatorFunc, TCreatorFunc, TDeletorFunc, TMutatorFunc}
+import com.evvo.island.population.{Minimize, Objective, Population, TObjective, TParetoFrontier}
 
 import scala.concurrent.{Await, Future}
 import scala.concurrent.duration.Duration
@@ -24,7 +26,6 @@ class EvvoIsland[Sol]
   private val creatorAgents = creators.map(c => CreatorAgent(c, pop))
   private val mutatorAgents = mutators.map(m => MutatorAgent(m, pop))
   private val deletorAgents = deletors.map(d => DeletorAgent(d, pop))
-
 
   override def runAsync(terminationCriteria: TTerminationCriteria)
   : Future[Unit] = {
@@ -72,5 +73,77 @@ class EvvoIsland[Sol]
     creatorAgents.foreach(_.stop())
     mutatorAgents.foreach(_.stop())
     deletorAgents.foreach(_.stop())
+  }
+}
+
+object EvvoIsland {
+  def builder[Sol](): EvvoIslandBuilder[Sol] = EvvoIslandBuilder[Sol]()
+}
+
+
+/**
+  * @param creators   the functions to be used for creating new solutions.
+  * @param mutators   the functions to be used for creating new solutions from current solutions.
+  * @param deletors   the functions to be used for deciding which solutions to delete.
+  * @param objectives the objective functions to maximize.
+  */
+case class EvvoIslandBuilder[Sol]
+(
+  creators: Set[TCreatorFunc[Sol]] = Set[TCreatorFunc[Sol]](),
+  mutators: Set[TMutatorFunc[Sol]] = Set[TMutatorFunc[Sol]](),
+  deletors: Set[TDeletorFunc[Sol]] = Set[TDeletorFunc[Sol]](),
+  objectives: Set[TObjective[Sol]] = Set[TObjective[Sol]]()
+) {
+  def addCreatorFromFunction(creatorFunc: CreatorFunctionType[Sol]): EvvoIslandBuilder[Sol] = {
+    this.copy(creators = creators + CreatorFunc(creatorFunc, creatorFunc.toString))
+  }
+
+  def addCreator(creatorFunc: TCreatorFunc[Sol]): EvvoIslandBuilder[Sol] = {
+    this.copy(creators = creators + creatorFunc)
+  }
+
+  def addMutatorFromFunction(mutatorFunc: MutatorFunctionType[Sol]): EvvoIslandBuilder[Sol] = {
+    this.copy(mutators = mutators + MutatorFunc(mutatorFunc, mutatorFunc.toString))
+  }
+
+  def addMutator(mutatorFunc: TMutatorFunc[Sol]): EvvoIslandBuilder[Sol] = {
+    this.copy(mutators = mutators + mutatorFunc)
+  }
+
+  def addDeletorFromFunction(deletorFunc: DeletorFunctionType[Sol]): EvvoIslandBuilder[Sol] = {
+    this.copy(deletors = deletors + DeletorFunc(deletorFunc, deletorFunc.toString))
+  }
+
+  def addDeletor(deletorFunc: TDeletorFunc[Sol]): EvvoIslandBuilder[Sol] = {
+    this.copy(deletors = deletors + deletorFunc)
+  }
+
+  // TODO this calling API is dangerous because it assumes minimization. it should be removed,
+  //      but this will require refactoring across examples
+  def addObjective(objective: ObjectiveFunctionType[Sol], name: String = "")
+  : EvvoIslandBuilder[Sol] = {
+    val realName = if (name == "") objective.toString() else name
+    this.copy(objectives = objectives + Objective(objective, realName, Minimize))
+  }
+
+  def addObjective(objective: Objective[Sol])
+  : EvvoIslandBuilder[Sol] = {
+    this.copy(objectives = objectives + objective)
+  }
+
+  def toProps()(implicit system: ActorSystem): Props = {
+    Props(new EvvoIslandActor[Sol](
+      creators.toVector,
+      mutators.toVector,
+      deletors.toVector,
+      objectives.toVector))
+  }
+
+  def buildLocalEvvo(): TEvolutionaryProcess[Sol] = {
+    new LocalEvvoIsland[Sol](
+      creators.toVector,
+      mutators.toVector,
+      deletors.toVector,
+      objectives.toVector)
   }
 }
